Use atexit/lazy initialization to manage lifetime
We use the start() and end() methods to control whether we are
collecting coverage information, and to trigger dumping that
information to a file. Additionally, we register an atexit
function that does the same, so that no matter how we are
exiting the module, the data will be properly dumped.

This is necessary because using Devel::QuickCover from our
Plack infrastructure is a bit more complex, in the presence
of calls to fork().

We have also checked that all memory is properly disposed of.