Add subroutine coverage (only for Pure-per subs)
It could be made to work with XS subs, but I don't think it's worth the effort.

The coverage is recorded by assuming that the first nexstate of a sub is always
executed, and is not inside a xscope/block/loop.

In the JSON output, there is a new "subs" object; the keys are in the form
"<sub name>,<first line>" and the values are the empty string if the subroutine
has not been covered or the phase in which the subroutine has been covered.